feat: 불필요한 step을 제거하고 인증 책임을 EmailVerificationStep에 위임합니다.
