+++ apps/web/src/features/apply/steps/EmailVerificationStep.tsx
@@ -1,35 +1,56 @@
-import { BlockButton } from "@ject/jds";
+import { BlockButton, Dialog } from "@ject/jds";
+import { useState } from "react";
 
+import { APPLY_DIALOG } from "@/constants/applyMessages.tsx";
 import { APPLY_TITLE } from "@/constants/applyPageData";
 import { ApplyStepLayout, AuthCodeForm } from "@/features/shared/components";
 import type { ApplyFunnelSteps } from "@/types/funnel";
 
+export type EmailVerificationEvents = {
+  onVerified: { email: string; authCode: string };
+  goToContinueWriting: undefined;
+  goBack: undefined;
+};
+
 interface EmailVerificationStepProps {
   context: ApplyFunnelSteps["이메일인증"];
-  onNext: (email: string, authCode: string) => void;
-  onExistingMember: (email: string) => void;
-  onBack: () => void;
+  dispatch: (
+    ...args:
+      | [type: "onVerified", payload: { email: string; authCode: string }]
+      | [type: "goToContinueWriting", payload?: undefined]
+      | [type: "goBack", payload?: undefined]
+  ) => void;
 }
 
 export function EmailVerificationStep({
   context,
-  onNext,
-  onExistingMember,
-  onBack,
+  dispatch,
 }: EmailVerificationStepProps) {
+  const [showExistingMemberDialog, setShowExistingMemberDialog] = useState(false);
+
+  const handleBack = () => dispatch("goBack");
+
+  const handleVerified = (email: string, authCode: string) => {
+    dispatch("onVerified", { email, authCode });
+  };
+
+  const handleExistingMember = () => {
+    setShowExistingMemberDialog(true);
+  };
+
   return (
     <ApplyStepLayout
       variant='apply'
       title={APPLY_TITLE.verifyEmail}
       current={0}
       jobFamily={context.jobFamily}
-      onBack={onBack}
+      onBack={handleBack}
     >
       <AuthCodeForm
         defaultEmail={context.email}
         sendGroupCode='AUTH_CODE'
-        onVerified={onNext}
-        onExistingMember={onExistingMember}
+        onVerified={handleVerified}
+        onExistingMember={handleExistingMember}
       />
       <BlockButton.Basic
         className='self-start'
@@ -41,6 +62,21 @@ export function EmailVerificationStep({
       >
         다음 단계로 진행하기
       </BlockButton.Basic>
+
+      <Dialog
+        open={showExistingMemberDialog}
+        onOpenChange={open => !open && setShowExistingMemberDialog(false)}
+        header={APPLY_DIALOG.tempSaved.header}
+        body={APPLY_DIALOG.tempSaved.body}
+        primaryAction={{
+          children: APPLY_DIALOG.tempSaved.primaryAction,
+          onClick: () => dispatch("goToContinueWriting"),
+        }}
+        secondaryAction={{
+          children: APPLY_DIALOG.tempSaved.secondaryAction,
+          onClick: () => setShowExistingMemberDialog(false),
+        }}
+      />
     </ApplyStepLayout>
   );
 }
