+++ apps/web/src/features/apply/steps/index.ts
@@ -1,10 +1,9 @@
 // ApplyFunnel (신규 유저)
-export { EmailVerificationStep } from "./EmailVerificationStep";
-export { PinSetupStep } from "./PinSetupStep";
 export {
-  ApplicationStatusStep,
-  type ApplicationStatusEvents,
-} from "./ApplicationStatusStep";
+  EmailVerificationStep,
+  type EmailVerificationEvents,
+} from "./EmailVerificationStep";
+export { PinSetupStep } from "./PinSetupStep";
 
 // ContinueWritingFunnel (이어쓰기)
 export {
